fix: compile error and generate region id by table_id and region number

+++ src/catalog/src/consts.rs
@@ -1,6 +1,6 @@
 pub const SYSTEM_CATALOG_NAME: &str = "system";
 pub const INFORMATION_SCHEMA_NAME: &str = "information_schema";
-pub const SYSTEM_CATALOG_TABLE_ID: u64 = 0;
+pub const SYSTEM_CATALOG_TABLE_ID: u32 = 0;
 pub const SYSTEM_CATALOG_TABLE_NAME: &str = "system_catalog";
 pub const DEFAULT_CATALOG_NAME: &str = "greptime";
 pub const DEFAULT_SCHEMA_NAME: &str = "public";
